DeleteSwipeWidget ist jetzt in einer eigenen Datei
